Implement extended output format with commit timestamp
With this change we introduce a new commandline option `--meta` to all
of our command line programs. This new option will change the output
format such that the commit timestamp of the latest commit included in
the prefetched github repository is included in a dedicated `meta`
section of the json output.